requirements.txt
pandas==2.1.4
pydantic==2.8.2
requests==2.32.3

duckdb==1.0.0
pytest==7.4.4
python-dotenv==1.0.1

# Providers
apache-airflow-providers-google
apache-airflow-providers-http
